// common/models/yahoo-finance/YahooFinanceCalendarEvents.ts
export interface YahooFinanceCalendarEvents {
  earningsDate: string | null;
  earningsCallDate: string | null;
  isEarningsDateEstimate: boolean | null;
  earningsAverage: string | null;
  earningsLow: string | null;
  earningsHigh: string | null;
  revenueAverage: string | null;
  revenueLow: string | null;
  revenueHigh: string | null;
  exDividendDate: string | null;
  dividendDate: string | null;
}